// apps/admin-web/src/.generated/api/model/operatorItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */

export interface OperatorItem {
  id: string;
  name: string;
  email: string;
  /** 할당된 역할 코드 */
  roleCode: string;
  /** 할당된 역할 표시명 */
  roleLabel: string;
  twoFactorEnabled: boolean;
  banned: boolean;
  /** @nullable */
  banReason?: string | null;
  /** @nullable */
  banExpires?: string | null;
  deleted: boolean;
  /** @nullable */
  deletedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}