// apps/admin-web/src/.generated/api/model/operatorsControllerGetOperatorOverviewV1200.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */
import type { OperatorOverviewResponse } from './operatorOverviewResponse';
import type { OperatorsControllerGetOperatorOverviewV1200Meta } from './operatorsControllerGetOperatorOverviewV1200Meta';

export type OperatorsControllerGetOperatorOverviewV1200 = {
  success: boolean;
  statusCode: number;
  path: string;
  requestId: string;
  timestamp: string;
  data: OperatorOverviewResponse;
  message?: string;
  meta?: OperatorsControllerGetOperatorOverviewV1200Meta;
};